Complete the HTML and CSS part of the Odin exercise
The HTML and CSS for the last two parts , the quote and
the call-to-action, have been added.
The Webpage has now been completed as far as the exercise as the
exercise goes.

In the future I will replace the non-sense texts with real information
and add a bit more functionality to the webpage.